chore(contracts): format

+++ packages/contracts/test/InitializeTest.sol
@@ -13,9 +13,7 @@ import {ClonesUpgradeable} from "oz-upgradeable/proxy/ClonesUpgradeable.sol";
 /// @author Federico Luzzi - <federico.luzzi@protonmail.com>
 contract InitializeTest is BaseTestSetup {
     function testZeroAddressKpiToken() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("ZeroAddressKpiToken()"));
         vm.prank(address(oraclesManager));
@@ -31,17 +29,11 @@ contract InitializeTest is BaseTestSetup {
     }
 
     function testZeroAddressArbitrator() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("ZeroAddressArbitrator()"));
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -49,30 +41,17 @@ contract InitializeTest is BaseTestSetup {
                 kpiToken: kpiToken,
                 templateId: _template.id,
                 templateVersion: _template.version,
-                data: abi.encode(
-                    address(0),
-                    0,
-                    "a",
-                    60,
-                    block.timestamp + 60,
-                    0
-                )
+                data: abi.encode(address(0), 0, "a", 60, block.timestamp + 60, 0)
             })
         );
     }
 
     function testEmptyQuestion() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("InvalidQuestion()"));
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -86,17 +65,11 @@ contract InitializeTest is BaseTestSetup {
     }
 
     function testZeroTimeout() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("InvalidQuestionTimeout()"));
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -110,17 +83,11 @@ contract InitializeTest is BaseTestSetup {
     }
 
     function testTooShortTimeout() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("InvalidQuestionTimeout()"));
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -128,30 +95,17 @@ contract InitializeTest is BaseTestSetup {
                 kpiToken: kpiToken,
                 templateId: _template.id,
                 templateVersion: _template.version,
-                data: abi.encode(
-                    address(1),
-                    0,
-                    "a",
-                    MINIMUM_QUESTION_TIMEOUT - 1,
-                    block.timestamp + 60,
-                    0
-                )
+                data: abi.encode(address(1), 0, "a", MINIMUM_QUESTION_TIMEOUT - 1, block.timestamp + 60, 0)
             })
         );
     }
 
     function testOpeningTimestampInThePast() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("InvalidOpeningTimestamp()"));
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -165,21 +119,17 @@ contract InitializeTest is BaseTestSetup {
     }
 
     function testOpeningTimestampMinimumAnswerWindows() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         vm.expectRevert(abi.encodeWithSignature("InvalidOpeningTimestamp()"));
         address kpiToken = address(1);
-        uint questionTimeout = 60;
-        uint openingTimestamp = block.timestamp + 2;
+        uint256 questionTimeout = 60;
+        uint256 openingTimestamp = block.timestamp + 2;
         vm.mockCall(
             kpiToken,
             abi.encodeWithSignature("expiration()"),
             // just a second before the value is valid
-            abi.encode(
-                uint256(questionTimeout * MINIMUM_ANSWER_WINDOWS + openingTimestamp - 1)
-            )
+            abi.encode(uint256(questionTimeout * MINIMUM_ANSWER_WINDOWS + openingTimestamp - 1))
         );
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
@@ -188,38 +138,23 @@ contract InitializeTest is BaseTestSetup {
                 kpiToken: kpiToken,
                 templateId: _template.id,
                 templateVersion: _template.version,
-                data: abi.encode(
-                    address(1),
-                    0,
-                    "a",
-                    questionTimeout,
-                    openingTimestamp,
-                    0
-                )
+                data: abi.encode(address(1), 0, "a", questionTimeout, openingTimestamp, 0)
             })
         );
     }
 
     function testSuccess() external {
-        RealityV3Oracle oracleInstance = RealityV3Oracle(
-            ClonesUpgradeable.clone(address(realityV3OracleTemplate))
-        );
+        RealityV3Oracle oracleInstance = RealityV3Oracle(ClonesUpgradeable.clone(address(realityV3OracleTemplate)));
         Template memory _template = oraclesManager.template(1);
         bytes32 _questionId = bytes32("questionId");
         vm.mockCall(
             REALITY_V3_ADDRESS,
-            abi.encodeWithSignature(
-                "askQuestionWithMinBond(uint256,string,address,uint32,uint32,uint256,uint256)"
-            ),
+            abi.encodeWithSignature("askQuestionWithMinBond(uint256,string,address,uint32,uint32,uint256,uint256)"),
             abi.encode(_questionId)
         );
         uint256 _openingTs = block.timestamp + 60;
         address kpiToken = address(1);
-        vm.mockCall(
-            kpiToken,
-            abi.encodeWithSignature("expiration()"),
-            abi.encode(2 ** 128)
-        );
+        vm.mockCall(kpiToken, abi.encodeWithSignature("expiration()"), abi.encode(2 ** 128));
         vm.prank(address(oraclesManager));
         oracleInstance.initialize(
             InitializeOracleParams({
@@ -233,27 +168,12 @@ contract InitializeTest is BaseTestSetup {
 
         assertEq(oracleInstance.template().id, _template.id);
 
-        vm.mockCall(
-            REALITY_V3_ADDRESS,
-            abi.encodeWithSignature("getArbitrator(bytes32)"),
-            abi.encode(address(1))
-        );
-        vm.mockCall(
-            REALITY_V3_ADDRESS,
-            abi.encodeWithSignature("getTimeout(bytes32)"),
-            abi.encode(uint32(60))
-        );
-        vm.mockCall(
-            REALITY_V3_ADDRESS,
-            abi.encodeWithSignature("getOpeningTS(bytes32)"),
-            abi.encode(_openingTs)
-        );
+        vm.mockCall(REALITY_V3_ADDRESS, abi.encodeWithSignature("getArbitrator(bytes32)"), abi.encode(address(1)));
+        vm.mockCall(REALITY_V3_ADDRESS, abi.encodeWithSignature("getTimeout(bytes32)"), abi.encode(uint32(60)));
+        vm.mockCall(REALITY_V3_ADDRESS, abi.encodeWithSignature("getOpeningTS(bytes32)"), abi.encode(_openingTs));
         bytes memory _data = oracleInstance.data();
-        (
-            address _onChainReality,
-            bytes32 _onChainQuestionId,
-            string memory _onChainQuestion
-        ) = abi.decode(_data, (address, bytes32, string));
+        (address _onChainReality, bytes32 _onChainQuestionId, string memory _onChainQuestion) =
+            abi.decode(_data, (address, bytes32, string));
         assertEq(_onChainReality, REALITY_V3_ADDRESS);
         assertEq(_onChainQuestionId, _questionId);
         assertEq(_onChainQuestion, "a");
